+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+
+## [1.6.0] - 2024-01-03
+
 ### Fixed
 
 - Fix an error in `Node.list_networks()` (Issue